requirements.txt
keras
tensorflow
Pillow
numpy
pyside6
qt-material
